Update DB models, generate instance on startup, and add singleton
