+++ app/build.gradle
@@ -41,6 +41,6 @@ dependencies {
 
     def room_version = "2.4.1"
 
-    implementation "androidx.room:room-runtime:$room_version"
-    annotationProcessor "androidx.room:room-compiler:$room_version"
+    implementation "androidx.data:data-runtime:$room_version"
+    annotationProcessor "androidx.data:data-compiler:$room_version"
 }
