+++ app/src/main/java/io/app/notebook/data/AppDatabase.java
@@ -1,0 +1,10 @@
+package io.app.notebook.data;
+
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+
+@Database(entities = {UserEntity.class, NoteEntity.class}, version = 1)
+public abstract class AppDatabase extends RoomDatabase {
+    public abstract UserDAO userDAO();
+    public abstract NoteDAO noteDAO();
+}
