+++ app/src/main/java/io/app/notebook/data/Data.kt
@@ -1,0 +1,8 @@
+package io.app.notebook.data
+
+import io.app.notebook.data.AppDatabase
+
+object Data {
+    lateinit var db: AppDatabase
+    var currentUserId = -1
+}
