+++ app/src/main/java/io/app/notebook/data/UserEntity.java
@@ -1,6 +1,5 @@
-package io.app.notebook.room;
+package io.app.notebook.data;
 
-import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
